Change IDs from Integer to Long.

+++ src/main/java/ru/kolaer/permit/component/EmptyObjects.java
@@ -12,9 +12,9 @@ import java.util.List;
 public final class EmptyObjects {
 
     public static final PostEntity DEFAULT_POST =
-            new PostEntity(-1, "", "", -1, "", Collections.emptyList()){
+            new PostEntity(-1L, "", "", -1, "", Collections.emptyList()){
                 @Override
-                public void setId(Integer id) {throw new UnsupportedOperationException("Post entity object is empty! You can't write value to empty objects.");}
+                public void setId(Long id) {throw new UnsupportedOperationException("Post entity object is empty! You can't write value to empty objects.");}
                 @Override
                 public void setName(String name) {throw new UnsupportedOperationException("Post entity object is empty! You can't write value to empty objects.");}
                 @Override
@@ -28,9 +28,9 @@ public final class EmptyObjects {
     };
 
     public static final DepartmentEntity DEFAULT_DEPARTMENT =
-            new DepartmentEntity(-1, "", "", Collections.emptyList()){
+            new DepartmentEntity(-1L, "", "", Collections.emptyList()){
                 @Override
-                public void setId(Integer id) {throw new UnsupportedOperationException("Department entity object is empty! You can't write value to empty objects.");}
+                public void setId(Long id) {throw new UnsupportedOperationException("Department entity object is empty! You can't write value to empty objects.");}
                 @Override
                 public void setName(String name) {throw new UnsupportedOperationException("Department entity object is empty! You can't write value to empty objects.");}
                 @Override
@@ -40,9 +40,9 @@ public final class EmptyObjects {
     };
 
     public static final EmployeeEntity DEFAULT_EMPLOYEE =
-            new EmployeeEntity(-1, "", new Date(0), -1, "", "", "", "", "", DEFAULT_DEPARTMENT, DEFAULT_POST){
+            new EmployeeEntity(-1L, "", new Date(0), -1, "", "", "", "", "", DEFAULT_DEPARTMENT, DEFAULT_POST){
                 @Override
-                public void setId(Integer id) {throw new UnsupportedOperationException("Employee entity object is empty! You can't write value to empty objects.");}
+                public void setId(Long id) {throw new UnsupportedOperationException("Employee entity object is empty! You can't write value to empty objects.");}
                 @Override
                 public void setInitials(String initials) {throw new UnsupportedOperationException("Employee entity object is empty! You can't write value to empty objects.");}
                 @Override
@@ -55,18 +55,18 @@ public final class EmptyObjects {
                 public void setPost(PostEntity post) {throw new UnsupportedOperationException("Employee entity object is empty! You can't write value to empty objects.");}
     };
 
-    public static final RoleEntity DEFAULT_ROLE = new RoleEntity(-1, DEFAULT_EMPLOYEE, "ROLE_USER") {
+    public static final RoleEntity DEFAULT_ROLE = new RoleEntity(-1L, DEFAULT_EMPLOYEE, "ROLE_USER") {
         @Override
-        public void setId(Integer id) {throw new UnsupportedOperationException("Role entity object is empty! You can't write value to empty objects.");}
+        public void setId(Long id) {throw new UnsupportedOperationException("Role entity object is empty! You can't write value to empty objects.");}
         @Override
         public void setRole(String role) {throw new UnsupportedOperationException("Role entity object is empty! You can't write value to empty objects.");}
         @Override
         public void setEmployee(EmployeeEntity employee) {throw new UnsupportedOperationException("Role entity object is empty! You can't write value to empty objects.");}
     };
 
-    public static final RoleEntity DEFAULT_FULL_ROLE = new RoleEntity(-1, DEFAULT_EMPLOYEE, "") {
+    public static final RoleEntity DEFAULT_FULL_ROLE = new RoleEntity(-1L, DEFAULT_EMPLOYEE, "") {
         @Override
-        public void setId(Integer id) {throw new UnsupportedOperationException("Full Role entity object is empty! You can't write value to empty objects.");}
+        public void setId(Long id) {throw new UnsupportedOperationException("Full Role entity object is empty! You can't write value to empty objects.");}
         @Override
         public void setRole(String role) {throw new UnsupportedOperationException("Full Role entity object is empty! You can't write value to empty objects.");}
         @Override
