+++ src/main/java/ru/kolaer/permit/dao/WorkEventDao.java
@@ -8,7 +8,7 @@ import java.util.List;
  * Created by Danilov on 01.06.2017.
  */
 public interface WorkEventDao extends BasePageDao<WorkEvent> {
-    List<WorkEvent> findByIdPermit(Integer idPermit, boolean findRemoved);
+    List<WorkEvent> findByIdPermit(Long idPermit, boolean findRemoved);
 
-    boolean deleteEmployee(Integer idWorkEvent, Integer idExecuter);
+    boolean deleteEmployee(Long idWorkEvent, Long idExecuter);
 }
