+++ src/main/java/ru/kolaer/permit/entity/PermitStatusHistoryEntity.java
@@ -29,7 +29,7 @@ public class PermitStatusHistoryEntity extends BaseEntity {
     private Date statusDate;
 
     @Column(name = "permit_id", insertable=false, updatable=false)
-    private Integer permitId;
+    private Long permitId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "permit_id", nullable = false)
