+++ src/main/java/ru/kolaer/permit/service/PermitConverter.java
@@ -4,5 +4,5 @@ package ru.kolaer.permit.service;
  * Created by danilovey on 06.06.2017.
  */
 public interface PermitConverter<T> {
-    T convertPermit(Integer idPermit);
+    T convertPermit(Long idPermit);
 }
